backend/requirements.txt
fastapi
uvicorn
apscheduler
python-dotenv
reportlab
pandas
google-generativeai
